// modules/transfer/modules/transfer-365-movil/enum/form-control.enum.ts
export enum EM365FormControlName {
  SOURCE_ACCOUNT = 'sourceAccount',
  BENEFICIARY = 'beneficiary',
  BANK = 'bank',
  IS_SHOW_FAVORITES = 'isShowFavorites',
  NUMBER_PHONE = 'numberPhone',
  AMOUNT = 'amount',
  COMMENT = 'comment',
}
export enum EM365FormControlNameEntered {
  NAME_BENEFICIARY = 'nameBeneficiary',
  EMAIL = 'email',
}


export enum EM365TypeLoadBeneficiary {
  REGISTERED = '1',
  ENTERED = '2',
}