// modules/transfer/modules/transfer-365-movil/enum/url-collection.enum.ts
export enum EM365UrlCollection {
  HOME = 'transfer/365-movil',
  CONFIRMATION = 'transfer/365-movil/confirmation',
  TERMS_CONDITION = 'transfer/365-movil/terms-condition',
  VOUCHER = 'transfer/365-movil/voucher',
  ST_DETAIL = 'transfer/365-movil/st/detail',
  ST_OPERATION = 'transfer/365-movil/st/operation',
  ST_MODIFY = 'transfer/365-movil/st/modify',
  ST_MODIFY_CONFIRM = 'transfer/365-movil/st/confirmation',
  ST_MODIFY_VOUCHER = 'transfer/365-movil/st/voucher',
}