// modules/transfer/modules/transfer-365-movil/enum/route-protected.enum.ts
export enum S365RouteProtected {
  HOME = 'm365:home',
  TERMS_CONDITION = 'm365:terms_condition',
  CONFIRMATION = 'm365:confirmation',
  VOUCHER = 'm365:voucher',
  ST_DETAIL = 'm365:st_detail',
  ST_MODIFY = 'm365:st_modify',
  ST_MODIFY_CONFIRM = 'm365:st_modify_confirm',
  ST_MODIFY_VOUCHER = 'm365:st_modify_voucher',
  ST_OPERATION = 'm365:st_operation',
}